feat: Refactor OTP feature by restructuring UI components and updating imports
