+++ lib/core/routing/app_router.dart
@@ -7,7 +7,7 @@ import 'package:nexus/features/layout/app_layout.dart';
 import '../../features/auth/forgot_password/ui/forgot_password_screen.dart';
 import '../../features/auth/login/logic/login_cubit.dart';
 import '../../features/auth/login/ui/login_screen.dart';
-import '../../features/auth/otp/presentation/otp_screen.dart';
+import '../../features/auth/otp/ui/otp_screen.dart';
 import '../../features/auth/reset_password/presentation/reset_password_screen.dart';
 import '../../features/auth/sign_up/ui/sign_up_screen.dart';
 import '../../features/burn_scan/burn_scan_screen.dart';
