+++ lib/features/auth/otp/ui/otp_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nexus/core/constants/assets.dart';
-import 'package:nexus/features/auth/otp/presentation/widgets/otp_body.dart';
+import 'package:nexus/features/auth/otp/ui/widgets/otp_body.dart';
 
 class OtpScreen extends StatelessWidget {
   const OtpScreen({super.key});
