+++ lib/features/auth/otp/ui/widgets/otp_body.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:nexus/core/widgets/custom_header.dart';
 import 'package:nexus/core/helpers/helper_methods/spacing.dart';
-import 'package:nexus/features/auth/otp/presentation/widgets/otp_form.dart';
-import 'package:nexus/features/auth/otp/presentation/widgets/resend_otp_text.dart';
+import 'package:nexus/features/auth/otp/ui/widgets/otp_form.dart';
+import 'package:nexus/features/auth/otp/ui/widgets/resend_otp_text.dart';
 
 class OtpBody extends StatelessWidget {
   const OtpBody({super.key});
@@ -17,7 +17,7 @@ class OtpBody extends StatelessWidget {
           const CustomHeader(
             text: 'Verification',
           ),
-          verticalSpace(240.h),
+          verticalSpace(240),
           const FractionallySizedBox(
             widthFactor: 1,
             child: OtpForm(),
